+++ components/ColdbrewIntro.jsx
@@ -1,0 +1,12 @@
+const ColdbrewIntro = () => {
+    
+
+    return (
+      <>
+
+      </>
+    )
+
+}
+
+export default ColdbrewIntro
